third_party/WebKit/LayoutTests/xr/requestDevice_one_device.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../vr/resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/vr_service.mojom.js"></script>
<script src="../vr/resources/mock-vr-service.js"></script>
<script src="../vr/resources/test-constants.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t) => {
  return navigator.xr.requestDevice().then( (device) => {
    t.step( () => {
      assert_true(device != null);
      assert_equals(device.deviceName, 'Google, Inc. Daydream View');
      assert_false(device.isExternal);
    }, "requestDevice returned correct results");
  }, (err) => {
    t.step( () => {
      assert_unreached("requestDevice rejected");
    });
  }).then( () => {
    t.done();
  });
}, [fakeDisplays["Pixel"]],
"navigator.xr.requestDevice properly returns a single device");

</script>
